simple-flask-app/requirements.txt
flask
flask-lambda-python36